#8 Use demand from solver and not from problem,
because the problems free is before the applieance of any
remove optimization events. The supply selectors gets executed
when the remove events are already created. Remove events are
not reverted inside the contrains based repairer and therefore
the problem's view of demands is outdated.

Instead the repair's own view is used.